pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-llama-sre-orchestrator-root"
version = "0.1.0"
description = "Repo-root wrapper for the llama_sre_orchestrator OpenEnv environment"
requires-python = ">=3.10"
dependencies = [
  "openenv-core[core]==0.2.3",
  "fastapi>=0.115.0",
  "uvicorn>=0.24.0",
]

[project.scripts]
server = "server.app:main"
